Refactor version fixtures
