+++ p2panda-rs/src/test_utils/fixtures/new_fixtures/mod.rs
@@ -11,6 +11,7 @@ mod hash_fixtures;
 mod identity_fixtures;
 mod operation_fixtures;
 mod schema_fixtures;
+mod version_fixtures;
 
 pub use document_fixtures::*;
 pub use entry_fixtures::*;
@@ -18,3 +19,4 @@ pub use hash_fixtures::*;
 pub use identity_fixtures::*;
 pub use operation_fixtures::*;
 pub use schema_fixtures::*;
+pub use version_fixtures::*;
