+++ p2panda-rs/src/test_utils/fixtures/new_fixtures/operation_fixtures.rs
@@ -183,3 +183,20 @@ pub fn meta_operation(
 pub fn operation_encoded(operation: Operation) -> OperationEncoded {
     OperationEncoded::try_from(&operation).unwrap()
 }
+
+/// Invalid YASMF hash in `document` with correct length but unknown hash format identifier.
+#[fixture]
+pub fn operation_encoded_invalid_relation_fields() -> OperationEncoded {
+    // {
+    //   "action": "create",
+    //   "schema": "venue_0020c65567ae37efea293e34a9c7d13f8f2bf23dbdc3b5c7b9ab46293111c48fc78b",
+    //   "version": 1,
+    //   "fields": {
+    //     "locations": {
+    //       "type": "relation",
+    //       "value": "83e2043738f2b5cdcd3b6cb0fbb82fe125905d0f75e16488a38d395ff5f9d5ea82b5"
+    //     }
+    //   }
+    // }
+    OperationEncoded::new("A466616374696F6E6663726561746566736368656D61784A76656E75655F30303230633635353637616533376566656132393365333461396337643133663866326266323364626463336235633762396162343632393331313163343866633738626776657273696F6E01666669656C6473A1696C6F636174696F6E73A264747970656872656C6174696F6E6576616C756578443833653230343337333866326235636463643362366362306662623832666531323539303564306637356531363438386133386433393566663566396435656138326235").unwrap()
+}
